@@ -3,7 +3,7 @@
   "name": "iflow/framework",
   "description": "iflowProject iflowFramework",
   "require": {
-    "php": ">=8.0.0",
+    "php": ">=8.1.0",
     "psr/log": "~1.0",
     "psr/container": "~1.0",
     "psr/simple-cache": "^1.0",
@@ -16,6 +16,7 @@
     "workerman/workerman": "^4.0",
     "webonyx/graphql-php": "^14.9",
     "iflow/template": "^0.0.1",
+    "iflow/router": "^0.0.1",
     "ext-json": "*",
     "ext-mbstring": "*",
     "ext-bcmath": "*",
@@ -35,7 +36,8 @@
   ],
   "autoload": {
     "psr-4": {
-      "iflow\\": "src"
+      "iflow\\": "src",
+      "app\\": "app"
     }
   },
   "autoload-dev": {
