@@ -3,9 +3,8 @@
 
 namespace iflow;
 
-
-use Exception;
 use iflow\contract\Arrayable;
+use ReturnTypeWillChange;
 use Traversable;
 
 class Collection implements \ArrayAccess, \Countable, \IteratorAggregate, \JsonSerializable, Arrayable
@@ -13,26 +12,23 @@ class Collection implements \ArrayAccess, \Countable, \IteratorAggregate, \JsonS
 
     protected array $items = [];
 
-    public function getIterator()
-    {
+    public function getIterator(): Traversable {
         // TODO: Implement getIterator() method.
         return new \ArrayIterator($this->items);
     }
 
-    public function offsetExists($offset)
-    {
+    public function offsetExists(mixed $offset): bool {
         // TODO: Implement offsetExists() method.
         return isset($this->items[$offset]) || !empty($this->items[$offset]);
     }
 
-    public function offsetGet($offset)
-    {
+    public function offsetGet(mixed $offset): mixed {
         // TODO: Implement offsetGet() method.
         return $this->offsetExists($offset) ? $this->items[$offset] : [];
     }
 
-    public function offsetSet($offset, $value)
-    {
+    #[ReturnTypeWillChange]
+    public function offsetSet(mixed $offset, mixed $value) {
         // TODO: Implement offsetSet() method.
         if (is_null($offset)) {
             return $this->items[] = $value;
@@ -40,38 +36,32 @@ class Collection implements \ArrayAccess, \Countable, \IteratorAggregate, \JsonS
         return $this->items[$offset] = $value;
     }
 
-    public function offsetUnset($offset)
-    {
+    public function offsetUnset(mixed $offset): void {
         // TODO: Implement offsetUnset() method.
         unset($this->items[$offset]);
     }
 
-    public function count()
-    {
+    public function count(): int {
         // TODO: Implement count() method.
         return count($this->items);
     }
 
-    public function jsonSerialize(): array
-    {
+    public function jsonSerialize(): array {
         // TODO: Implement jsonSerialize() method.
         return $this->toArray();
     }
 
-    public function toArray(): array
-    {
+    public function toArray(): array {
         return array_map(function ($value) {
             return $value instanceof Arrayable ? $value->toArray() : $value;
         }, $this->items);
     }
 
-    public function toJson(int $options = JSON_UNESCAPED_UNICODE): string
-    {
+    public function toJson(int $options = JSON_UNESCAPED_UNICODE): string {
         return json_encode($this->toArray(), $options);
     }
 
-    public function all(): array
-    {
+    public function all(): array {
         return $this->items;
     }
 }
